feat(monetization): a settled day can be corrected, and a rupee cannot vanish
Three things the ledger could not do.

It could not correct itself. A settled day was never re-settled, an
earning row's money columns were the only record, and the words
adjustment and reversed existed solely as enum values no code wrote.
There are now two primitives, both append-only. PostAdjustment writes a
signed transaction keyed on its cause so the same cause applied twice
moves money once, and mirrors it into the double-entry ledger.
ReverseFundEarning marks a row reversed and, if it had been credited,
posts the opposing net through that primitive and backs the platform fee
leg out too, so reversing an earning does not leave the platform holding
a fee on income that no longer exists. A settled row's money columns are
never updated. A reversed row keeps its unique slot on purpose; a
corrected re-accrual is a new adjustment, not a second row.

It could lose money to rounding. One short-video view is 0.3 paise of
base gross and integer truncation per row discarded it, so a small
creator's flick views could sum to nothing. Accrual is now in
micro-paise with a per-creator carry; truncation happens once per day at
the carry boundary and the remainder rolls forward, so the sum of daily
paise plus the final carry equals the exact total. One view a day for
four days now yields a paise on the fourth. Every row records the rate
and band it was priced with, a rule version, and a hash of its inputs so
a changed input is detected rather than silently ignored. The hash is
over values only, because the analytics rollup rewrites open days inside
its window and a timestamp in the hash would have made settlement refuse
identical data.

It had no budget. The founder's decision is a prospective stop: a
per-period cap, accrual halts for the rest of the period once reached,
and nothing already earned is reduced. Skipped days are written as zero
rows with a reason rather than left missing, creators accrue in a fixed
order so who hits the cap is deterministic, lowering a cap below what is
already accrued is refused, and a cadence change is refused while a
period holds accruals. The lock order is budget, then carry, then
earnings, and a concurrency test pins it: six creators at once against a
cap that fits two, and the cap holds exactly.

One correction to the plan this implements. Its migration guard converted
decimal columns to integers by multiplying by a hundred. The live
database's columns are decimal and already hold integer paise written by
Go, so that would have inflated every balance a hundredfold. The guard
now casts, refuses on any fractional value, and is proven on both
database shapes.

Every accrual skip is now counted by reason and logged, instead of a
silent continue. Two admin routes exist for the reversal and the budget,
behind the admin scope and, for now, behind the beta boundary.

Verified: vet, build, unit and integration suites green against a scratch
database; the lock order, the budget serialisation under concurrency, the
credited-reversal branch and the fee leg each fail their test when
neutered. The January remediation is written as a runbook and not run.

+++ Architecture/services/monetization-service/database/setup.sql
@@ -3,11 +3,18 @@
 -- creator_ledger (renamed from `wallets` 2026-04-30, Phase 2 §D4).
 -- Holds creator earnings: lifetime_earnings, pending_payout, balance.
 -- NOT a consumer wallet — that lives in wallet-service.
+--
+-- Money columns are BIGINT paise. They were declared DECIMAL(12,2) here
+-- until 2026-09-11 while migration 005 only converted them when `wallets`
+-- was still a base table, so every install made after migration 012 got
+-- NUMERIC columns that the Go code wrote int64 paise into. Migration 018
+-- converts those in place (a cast, not a scaling); this declaration is
+-- what a fresh install gets.
 CREATE TABLE IF NOT EXISTS creator_ledger (
     user_id       UUID PRIMARY KEY,
-    balance       DECIMAL(12,2) NOT NULL DEFAULT 0.00,
-    lifetime_earnings DECIMAL(12,2) NOT NULL DEFAULT 0.00,
-    pending_payout DECIMAL(12,2) NOT NULL DEFAULT 0.00,
+    balance       BIGINT NOT NULL DEFAULT 0,
+    lifetime_earnings BIGINT NOT NULL DEFAULT 0,
+    pending_payout BIGINT NOT NULL DEFAULT 0,
     currency      TEXT NOT NULL DEFAULT 'INR',
     is_frozen     BOOLEAN NOT NULL DEFAULT FALSE,
     created_at    TIMESTAMPTZ NOT NULL DEFAULT now(),
